Rename source files and fix extend
