+++ CS2Fixes.vcxproj
@@ -181,7 +181,7 @@
     <ClCompile Include="src\mempatch.cpp" />
     <ClCompile Include="src\patches.cpp" />
     <ClCompile Include="src\playermanager.cpp" />
-	<ClCompile Include="src\rockthevote.cpp" />
+	<ClCompile Include="src\votemanager.cpp" />
     <ClCompile Include="sdk\tier1\convar.cpp" />
     <ClCompile Include="src\utils\entity.cpp" />
     <ClCompile Include="src\utils\plat_unix.cpp" />
@@ -216,7 +216,7 @@
     <ClInclude Include="src\addresses.h" />
     <ClInclude Include="src\playermanager.h" />
     <ClInclude Include="src\recipientfilters.h" />
-	<ClInclude Include="src\rockthevote.h" />
+	<ClInclude Include="src\votemanager.h" />
     <ClInclude Include="src\utils\entity.h" />
     <ClInclude Include="src\utils\module.h" />
     <ClInclude Include="src\utils\plat.h" />
