+++ CS2Fixes.vcxproj.filters
@@ -98,7 +98,7 @@
     <ClCompile Include="src\httpmanager.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-	<ClCompile Include="src\rockthevote.cpp">
+	<ClCompile Include="src\votemanager.cpp">
 		<Filter>Source Files</Filter>
 	</ClCompile>
   </ItemGroup>
@@ -199,7 +199,7 @@
     <ClInclude Include="src\httpmanager.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-	<ClInclude Include="src\rockthevote.h">
+	<ClInclude Include="src\votemanager.h">
 		<Filter>Header Files</Filter>
 	</ClInclude>
   </ItemGroup>
