+++ src/votemanager.h
